REFACTOR: Renamed component
